backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Document Processing
pdfplumber==0.10.3
python-docx==1.1.0
PyPDF2==3.0.1

# Data Processing
python-dateutil==2.8.2
pydantic==2.5.3
pydantic-settings==2.1.0

# NER Models (Work Item 3)
spacy==3.7.2

# Utilities
aiofiles==23.2.1
python-magic==0.4.27

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

